auto extract
In this code, we use the urlparse function from the urllib.parse module to extract the scheme, netloc, and path components of the full URL. We then use os.path.splitext to split the file extension from the file name and the path. Finally, we construct the file_name_format string using the os.path.splitext results and the %d format specifier.